[rest] add rest implementation
Using Oat++ framework and kudu client a controller has been
generated to map the requests to respective endpoints. To avoid
potential issues due to rest server receiving requests before master
servers are up, a client is generated to serve the each individual
request. Data Transfer Objects of Oat++ framework have been introduced
to correctly display the required JSON Schemas in the Swagger component
and to parse the incoming/outgoing data in the correct format. Running
the server will block the calling thread.

As the server uses a kudu client for the requests, and kudu client needs
to know the addresses of the master servers, the better choice was to
embed into the master, instead of having a seperate server. This also
has the benefit of keeping the endpoint alive in case one of the master
servers stop. The rest endpoint is off by default and requires a bind
adress(default 0.0.0.0:8061), which can be provided by flags. A seperate
threadpool takes care of running the rest endpoint, not to block the
master server.

An integration test is included, which checks all of the endpoints by
sending a request to them and asserting the incoming response. The bind
address for the rest is set dynamically at each test case.

Change-Id: I2ca3121fd7e95a1267853be45cb5f5855298c763