KUDU-3408 Better "kudurc" documentation
Kudu CLI provides the possibility to have cluster aliases to provide the
list of master addresses in a "kudurc" file. This is quite useful
feature, as it negates the need for the users to remember/type the list
of master addresses everytime they want to invoke a Kudu CLI command.
This patch adds a reference to this feature in Quickstart.adoc to make
it more apparent to new users. It also makes the possibility of using
environment variables for substituting Master Addresses with KUDU CLI
more apparent.

Change-Id: I33d9a6d579343ff7333af955bf309ac66bfb0c7b